+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added `formatFn` for Formatters and `regex`, `flags`, and `function` for Filters to ease `intel.config()`
 - added `Logger#trace` and `intel.TRACE` level
 - added `exception: boolean` and `uncaughtException: boolean` to LogRecord
+- added `Logger.NONE` and `Logger.ALL` levels
 - changed `intel.config` to remove default ROOT console handler
 
 ## v0.4.0 0 2013-12-04